requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
duckduckgo-search==7.4.4
fastapi==0.115.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
openai==1.63.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
postgrest==0.19.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.0
requests==2.32.3
rich==13.9.4
six==1.17.0
smolagents==1.9.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
